+++ src/tests/unit/test_seqbox_utils_project.py
@@ -6,7 +6,6 @@ sys.path.append('../')
 sys.path.append('./')
 sys.path.append('../scripts')
 from app import app, db
-from scripts.utils.check import check_project
 from scripts.utils.db import add_project, query_projects, add_group
 
 
@@ -77,14 +76,3 @@ class TestSeqboxUtilsProjects(TestCase):
         project_info = {'project_name': 'Testproject', 'group_name': 'TestGroup', 'institution': 'Test Institution'}
         with self.assertRaises(SystemExit, msg='Missing project_details') as cm:
             add_project(project_info)
-
-    def test_check_project(self):
-        self.assertEqual(None, check_project({'project_name': 'Testproject', 'group_name': 'TestGroup', 'institution': 'Test Institution', 'project_details': 'sequencing'}))
-
-        # empty
-        with self.assertRaises(SystemExit) as cm:
-            check_project({'project_name': '', 'group_name': 'TestGroup', 'institution': 'Test Institution', 'project_details': 'sequencing'})
-        with self.assertRaises(SystemExit) as cm:
-            check_project({'project_name': 'Testproject', 'group_name': '', 'institution': 'Test Institution', 'project_details': 'sequencing'})
-        with self.assertRaises(SystemExit) as cm:
-            check_project({'project_name': 'Testproject', 'group_name': 'TestGroup', 'institution': '', 'project_details': 'sequencing'})
